+++ node/src/test/kotlin/net/corda/node/services/statemachine/FlowFrameworkPersistenceTests.kt
@@ -3,7 +3,6 @@ package net.corda.node.services.statemachine
 import net.corda.core.crypto.random63BitValue
 import net.corda.core.flows.FlowLogic
 import net.corda.core.identity.Party
-import net.corda.core.serialization.internal.CheckpointSerializationDefaults
 import net.corda.core.utilities.getOrThrow
 import net.corda.node.services.persistence.checkpoints
 import net.corda.testing.core.ALICE_NAME
@@ -21,7 +20,6 @@ import org.junit.Ignore
 import org.junit.Test
 import rx.Observable
 import java.util.*
-import kotlin.streams.toList
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
@@ -148,32 +146,6 @@ class FlowFrameworkPersistenceTests {
         assertEquals(payload2 + 1, firstAgain.receivedPayload2, "Received payload does not match the expected second value on Node 3")
         assertEquals(payload, secondFlow.getOrThrow().receivedPayload, "Received payload does not match the (restarted) first value on Node 2")
         assertEquals(payload + 1, secondFlow.getOrThrow().receivedPayload2, "Received payload does not match the expected second value on Node 2")
-    }
-
-    @Test
-    fun `Checkpoint status is being changed to RUNNABLE after suspension`() {
-        SuspendingFlow.hookBeforeCheckpoint = {
-            val flowFiber = (this as? FlowStateMachineImpl<*>)
-
-            // manually change the -in memory- Checkpoint.status to FAILED
-            val newState = flowFiber!!.transientState!!.value.copy(
-                   checkpoint =  flowFiber.transientState!!.value.checkpoint.copy(
-                        status = Checkpoint.FlowStatus.FAILED
-                   )
-            )
-            flowFiber.transientState = TransientReference(newState)
-        }
-        SuspendingFlow.hookAfterCheckpoint = {
-            // assert checkpoint is changed to RUNNABLE
-            val persistedCheckpoint = aliceNode.internals.checkpointStorage.getAllCheckpoints().toList().single()
-            assertEquals(Checkpoint.FlowStatus.RUNNABLE, persistedCheckpoint.second.status)
-        }
-
-        aliceNode.services.startFlow(SuspendingFlow()).resultFuture.getOrThrow()
-
-        // clear SendFlow companion object
-        SuspendingFlow.hookBeforeCheckpoint = {}
-        SuspendingFlow.hookAfterCheckpoint = {}
     }
 
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////
